FitMind/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
firebase-admin==6.5.0
python-dotenv==1.0.1
pydantic==2.9.2
google-genai>=0.5.0
gunicorn==23.0.0
stripe>=7.0.0
python-jose[cryptography]>=3.3.0
tzdata                        # Timezone databáza pre Slovakia (zoneinfo)
